@@ -22,7 +22,7 @@ config.plugins = [
 config.module = {
 	loaders: [
 		{include: /\.json$/, loaders: ["json-loader"]},
-		{include: /\.js$/, loaders: ["react-hot", "babel-loader?experimental&optional=runtime"], exclude: /(node_modules|lib)/}
+		{include: /\.js$/, loaders: ["react-hot", "babel-loader?stage=1&optional=runtime"], exclude: /(node_modules|lib)/}
 	]
 };
 
