@@ -22,7 +22,7 @@ module.exports = {
 	module:  {
 		loaders: [
 			{include: /\.json$/, loaders: ["json-loader"]},
-			{include: /\.js$/, loaders: ["babel-loader?experimental&optional=runtime"], exclude: /(node_modules|lib)/}
+			{include: /\.js$/, loaders: ["babel-loader?stage=1&optional=runtime"], exclude: /(node_modules|lib)/}
 		]
 	},
 	resolve: {
